Rename: api 인스턴스 이름 변경 -> rootApiInstance
